format metric code

See merge request megasearch/milvus!629

+++ cpp/src/metrics/PrometheusMetrics.cpp
@@ -16,12 +16,14 @@
 // under the License.
 
 
+#include "metrics/PrometheusMetrics.h"
 #include "cache/GpuCacheMgr.h"
-#include "PrometheusMetrics.h"
 #include "server/Config.h"
 #include "utils/Log.h"
 #include "SystemInfo.h"
 
+#include <string>
+#include <utility>
 
 namespace zilliz {
 namespace milvus {
@@ -47,93 +49,96 @@ PrometheusMetrics::Init() {
 
         // Exposer Registry
         exposer_ptr_->RegisterCollectable(registry_);
-    } catch (std::exception& ex) {
+    } catch (std::exception &ex) {
         SERVER_LOG_ERROR << "Failed to connect prometheus server: " << std::string(ex.what());
         return SERVER_UNEXPECTED_ERROR;
     }
 
     return SERVER_SUCCESS;
-
 }
 
-
 void
-PrometheusMetrics::CPUUsagePercentSet()  {
-    if(!startup_) return ;
+PrometheusMetrics::CPUUsagePercentSet() {
+    if (!startup_) return;
     double usage_percent = server::SystemInfo::GetInstance().CPUPercent();
     CPU_usage_percent_.Set(usage_percent);
 }
 
 void
 PrometheusMetrics::RAMUsagePercentSet() {
-    if(!startup_) return ;
+    if (!startup_) return;
     double usage_percent = server::SystemInfo::GetInstance().MemoryPercent();
     RAM_usage_percent_.Set(usage_percent);
 }
 
 void
 PrometheusMetrics::GPUPercentGaugeSet() {
-    if(!startup_) return;
+    if (!startup_) return;
     int numDevice = server::SystemInfo::GetInstance().num_device();
-    std::vector<unsigned long long > used_total = server::SystemInfo::GetInstance().GPUMemoryTotal();
-    std::vector<unsigned long long > used_memory = server::SystemInfo::GetInstance().GPUMemoryUsed();
+    std::vector<uint64_t> used_total = server::SystemInfo::GetInstance().GPUMemoryTotal();
+    std::vector<uint64_t> used_memory = server::SystemInfo::GetInstance().GPUMemoryUsed();
 
     for (int i = 0; i < numDevice; ++i) {
         prometheus::Gauge &GPU_percent = GPU_percent_.Add({{"DeviceNum", std::to_string(i)}});
-        double percent = (double)used_memory[i] / (double)used_total[i];
+        double percent = (double) used_memory[i] / (double) used_total[i];
         GPU_percent.Set(percent * 100);
     }
 }
 
-void PrometheusMetrics::GPUMemoryUsageGaugeSet() {
-    if(!startup_) return;
-    std::vector<unsigned long long> values = server::SystemInfo::GetInstance().GPUMemoryUsed();
-    constexpr unsigned long long MtoB = 1024*1024;
+void
+PrometheusMetrics::GPUMemoryUsageGaugeSet() {
+    if (!startup_) return;
+    std::vector<uint64_t> values = server::SystemInfo::GetInstance().GPUMemoryUsed();
+    constexpr uint64_t MtoB = 1024 * 1024;
     int numDevice = server::SystemInfo::GetInstance().num_device();
 
     for (int i = 0; i < numDevice; ++i) {
         prometheus::Gauge &GPU_memory = GPU_memory_usage_.Add({{"DeviceNum", std::to_string(i)}});
         GPU_memory.Set(values[i] / MtoB);
     }
-
 }
-void PrometheusMetrics::AddVectorsPerSecondGaugeSet(int num_vector, int dim, double time) {
+
+void
+PrometheusMetrics::AddVectorsPerSecondGaugeSet(int num_vector, int dim, double time) {
     // MB/s
-    if(!startup_) return;
+    if (!startup_) return;
 
-    long long MtoB = 1024*1024;
-    long long size = num_vector * dim * 4;
-    add_vectors_per_second_gauge_.Set(size/time/MtoB);
-
+    int64_t MtoB = 1024 * 1024;
+    int64_t size = num_vector * dim * 4;
+    add_vectors_per_second_gauge_.Set(size / time / MtoB);
 }
-void PrometheusMetrics::QueryIndexTypePerSecondSet(std::string type, double value) {
-    if(!startup_) return;
-    if(type == "IVF"){
+
+void
+PrometheusMetrics::QueryIndexTypePerSecondSet(std::string type, double value) {
+    if (!startup_) return;
+    if (type == "IVF") {
         query_index_IVF_type_per_second_gauge_.Set(value);
-    } else if(type == "IDMap"){
+    } else if (type == "IDMap") {
         query_index_IDMAP_type_per_second_gauge_.Set(value);
     }
-
 }
 
-void PrometheusMetrics::ConnectionGaugeIncrement() {
-    if(!startup_) return;
+void
+PrometheusMetrics::ConnectionGaugeIncrement() {
+    if (!startup_) return;
     connection_gauge_.Increment();
 }
 
-void PrometheusMetrics::ConnectionGaugeDecrement() {
-    if(!startup_) return;
+void
+PrometheusMetrics::ConnectionGaugeDecrement() {
+    if (!startup_) return;
     connection_gauge_.Decrement();
 }
 
-void PrometheusMetrics::OctetsSet() {
-    if(!startup_) return;
+void
+PrometheusMetrics::OctetsSet() {
+    if (!startup_) return;
 
     // get old stats and reset them
-    unsigned long long old_inoctets = SystemInfo::GetInstance().get_inoctets();
-    unsigned long long old_outoctets = SystemInfo::GetInstance().get_octets();
+    uint64_t old_inoctets = SystemInfo::GetInstance().get_inoctets();
+    uint64_t old_outoctets = SystemInfo::GetInstance().get_octets();
     auto old_time = SystemInfo::GetInstance().get_nettime();
-    std::pair<unsigned long long, unsigned long long> in_and_out_octets = SystemInfo::GetInstance().Octets();
+    std::pair<uint64_t, uint64_t> in_and_out_octets = SystemInfo::GetInstance().Octets();
     SystemInfo::GetInstance().set_inoctets(in_and_out_octets.first);
     SystemInfo::GetInstance().set_outoctets(in_and_out_octets.second);
     SystemInfo::GetInstance().set_nettime();
@@ -142,13 +147,14 @@ void PrometheusMetrics::OctetsSet() {
     constexpr double micro_to_second = 1e-6;
     auto now_time = std::chrono::system_clock::now();
     auto total_microsecond = METRICS_MICROSECONDS(old_time, now_time);
-    auto total_second = total_microsecond*micro_to_second;
-    if(total_second == 0) return;
-    inoctets_gauge_.Set((in_and_out_octets.first-old_inoctets)/total_second);
-    outoctets_gauge_.Set((in_and_out_octets.second-old_outoctets)/total_second);
+    auto total_second = total_microsecond * micro_to_second;
+    if (total_second == 0) return;
+    inoctets_gauge_.Set((in_and_out_octets.first - old_inoctets) / total_second);
+    outoctets_gauge_.Set((in_and_out_octets.second - old_outoctets) / total_second);
 }
 
-void PrometheusMetrics::CPUCoreUsagePercentSet() {
+void
+PrometheusMetrics::CPUCoreUsagePercentSet() {
     if (!startup_)
         return;
 
@@ -160,11 +166,12 @@ void PrometheusMetrics::CPUCoreUsagePercentSet() {
     }
 }
 
-void PrometheusMetrics::GPUTemperature() {
+void
+PrometheusMetrics::GPUTemperature() {
     if (!startup_)
         return;
 
-    std::vector<unsigned int> GPU_temperatures = server::SystemInfo::GetInstance().GPUTemperature();
+    std::vector<uint64_t> GPU_temperatures = server::SystemInfo::GetInstance().GPUTemperature();
 
     for (int i = 0; i < GPU_temperatures.size(); ++i) {
         prometheus::Gauge &gpu_temp = GPU_temperature_.Add({{"GPU", std::to_string(i)}});
@@ -172,7 +179,8 @@ void PrometheusMetrics::GPUTemperature() {
     }
 }
 
-void PrometheusMetrics::CPUTemperature() {
+void
+PrometheusMetrics::CPUTemperature() {
     if (!startup_)
         return;
 
@@ -184,7 +192,8 @@ void PrometheusMetrics::CPUTemperature() {
     }
 }
 
-void PrometheusMetrics::GpuCacheUsageGaugeSet() {
+void
+PrometheusMetrics::GpuCacheUsageGaugeSet() {
 //    std::vector<uint64_t > gpu_ids = {0};
 //    for(auto i = 0; i < gpu_ids.size(); ++i) {
 //        uint64_t cache_usage = cache::GpuCacheMgr::GetInstance(gpu_ids[i])->CacheUsage();
@@ -194,6 +203,6 @@ void PrometheusMetrics::GpuCacheUsageGaugeSet() {
 //    }
 }
 
-}
-}
-}
+} // namespace server
+} // namespace milvus
+} // namespace zilliz
